new directories structure for interfaces
