+++ includes/interfaces/ValidateInterface/IValidateFunction.hpp
@@ -1,0 +1,11 @@
+#ifndef IVALIDATE_FUNCTION_H
+# define IVALIDATE_FUNCTION_H
+
+class IValidateFunction
+{
+    public:
+        virtual bool operator()(std::string &) const = 0;
+        virtual ~IValidateFunction() {}
+};
+
+#endif
